+++ src/libs/types.ts
@@ -1,16 +1,29 @@
-import {BaseSchema} from "./BaseSchema";
+import {BaseSchema} from "../schemaClasses/BaseSchema";
 
-class AS<T> extends BaseSchema<T[]>
-{}
-class OS<S extends SchemaObject> extends BaseSchema<ObjectTypeOf<S>>
-{}
+import {ArraySchema} from "../schemaClasses/ArraySchema";
+import {BooleanSchema} from "../schemaClasses/BooleanSchema";
+import {EmailSchema} from "../schemaClasses/EmailSchema";
+import {NumberSchema} from "../schemaClasses/NumberSchema";
+import {NumericStringSchema} from "../schemaClasses/NumericStringSchema";
+import {ObjectSchema} from "../schemaClasses/ObjectSchema";
+import {StringSchema} from "../schemaClasses/StringSchema";
+
 export type ObjectTypeOf<S extends SchemaObject> = {
 	[K in keyof S]:
-		S[K] extends AS<infer T> ? T[] :
-		S[K] extends OS<infer S2> ? ObjectTypeOf<S2> :
-		S[K] extends BaseSchema<boolean> ? boolean :
-		S[K] extends BaseSchema<number> ? number :
-		S[K] extends BaseSchema<string> ? string :
+		S[K] extends ArraySchema<infer T> ? T[] :
+		S[K] extends ArraySchema<infer T, null> ? T[] | null :
+		S[K] extends ObjectSchema<infer S2> ? ObjectTypeOf<S2> :
+		S[K] extends ObjectSchema<infer S2, null> ? ObjectTypeOf<S2> | null :
+		S[K] extends BooleanSchema ? boolean :
+		S[K] extends BooleanSchema<null> ? boolean | null :
+		S[K] extends NumberSchema ? number :
+		S[K] extends NumberSchema<null> ? number | null :
+		S[K] extends StringSchema ? string :
+		S[K] extends StringSchema<null> ? string | null :
+		S[K] extends EmailSchema ? string :
+		S[K] extends EmailSchema<null> ? string | null :
+		S[K] extends NumericStringSchema ? string :
+		S[K] extends NumericStringSchema<null> ? string | null :
 		never
 }
 
@@ -24,7 +37,7 @@ export type Key = string | number;
 export interface Values<T = unknown>
 {
 	input: unknown;
-	output: T | null;
+	output: T;
 }
 
 /**